7-puts_half:  A function that puts_half - Prints the second half of a string.

+++ 0x05-pointers_arrays_strings/s_length.c
@@ -1,8 +1,8 @@
 #include"main.h"
 
-size_t s_length(const char *s)
+int s_length(const char *s)
 {
-    int length = 0; 
+    size_t length = 0; 
 
     while(s[length] != '\0')
     {
